feat: não bloquear admins

+++ app/Console/Commands/UpdateLqxBalances.php
@@ -46,7 +46,7 @@ class UpdateLqxBalances extends Command
      */
     public function handle()
     {
-        $output = new \Symfony\Component\Console\Output\ConsoleOutput();
+//        $output = new \Symfony\Component\Console\Output\ConsoleOutput();
 
         try {
             $wallets = UserWallet::with(['user', 'coin'])
@@ -61,38 +61,44 @@ class UpdateLqxBalances extends Command
                 $transactionController = new TransactionsController(new BalanceService());
                 $computed = $transactionController->balanceVerify($wallet->user->email);
 
-                $output->writeln("<info>-----------------------------</info>");
-                $output->writeln("<info>{$wallet->user->email}</info>");
-                $output->writeln("<info>local balance: {$wallet->balance}</info>");
-                $output->writeln("<info>verify balance: {$computed['balances']['LQX']['balance_computed']->balance}</info>");
-                $output->writeln("<info>verify sum transactions: {$computed['balances']['LQX']['balance']}</info>");
+//                $output->writeln("<info>-----------------------------</info>");
+//                $output->writeln("<info>{$wallet->user->email}</info>");
+//                $output->writeln("<info>local balance: {$wallet->balance}</info>");
+//                $output->writeln("<info>verify balance: {$computed['balances']['LQX']['balance_computed']->balance}</info>");
+//                $output->writeln("<info>verify sum transactions: {$computed['balances']['LQX']['balance']}</info>");
 
                 if ($computed['balances']['LQX']['balance'] < -0.001) {
-
-//                    if (!$wallet->user->is_under_analysis) {
-//                        $message = env("APP_NAME") . " - Usuário bloqueado: " . env("ADMIN_URL") . "/user/analysis/" . $wallet->user->email;
-//                        Mail::to(env('DEV_MAIL', 'cristianovelkan@gmail.com'))->send(new AlertsMail($message));
-//                        sleep(2);
-//                    }
-
                     $user = User::find($wallet->user_id);
-                    $user->is_under_analysis = true;
-                    $user->save();
 
-                    $user->tokens()->each(function ($token) {
-                        $token->delete();
-                    });
+                    if (!$user->is_admin) {
+                        if (!$user->is_under_analysis) {
 
-                    $output->writeln("<info>BLOQUEADO ANALISE</info>");
+                            $user->is_under_analysis = true;
+                            $user->save();
+
+                            $user->tokens()->each(function ($token) {
+                                $token->delete();
+                            });
+
+                            $message = env("APP_NAME") . " - Usuário bloqueado: " . env("ADMIN_URL") . "/user/analysis/" . $wallet->user->email;
+                            Mail::to(env('DEV_MAIL', 'cristianovelkan@gmail.com'))->send(new AlertsMail($message));
+                            sleep(2);
+                        }
+                    }
+
+//                    $output->writeln("<info>BLOQUEADO ANALISE</info>");
                 }
 
-                $wallet->balance = $computed['balances']['LQX']['balance'];
-                $wallet->save();
+                if (!$wallet->user->is_admin) {
+                    $wallet->balance = $computed['balances']['LQX']['balance'];
+                    $wallet->save();
+                }
             }
 
         } catch (\Exception $e) {
-            $output->writeln("<info>{$e->getMessage()}</info>");
-            $output->writeln("<info>{$e->getLine()} - {$e->getFile()}</info>");
+            throw new \Exception($e->getLine() . ' - ' . $e->getFile() . ' - '. $e->getMessage());
+//            $output->writeln("<info>{$e->getMessage()}</info>");
+//            $output->writeln("<info>{$e->getLine()} - {$e->getFile()}</info>");
         }
     }
 }
